feat: rate can be up to two dp

+++ src-tauri/migrations/2024-02-04-082255_create_line_items/up.sql
@@ -2,7 +2,7 @@
 CREATE TABLE line_items (
     id INTEGER PRIMARY KEY NOT NULL,
     name TEXT NOT NULL,
-    rate INTEGER NOT NULL,
+    rate REAL NOT NULL,
     customer_id INTEGER NOT NULL,
     FOREIGN KEY (customer_id)
         REFERENCES customers(id)
